Add Dynamo.Logging namespace.
