+++ src/DynamoCore/DSEngine/AstBuilder.cs
@@ -6,6 +6,7 @@ using System.Linq;
 using Dynamo.Interfaces;
 using Dynamo.Library;
 using Dynamo.Models;
+using Dynamo.Logging;
 
 using ProtoCore;
 using ProtoCore.AST.AssociativeAST;
