+++ src/DynamoCore/Models/NodeFactory.cs
@@ -5,6 +5,7 @@ using Dynamo.Interfaces;
 using Dynamo.Models.NodeLoaders;
 using Dynamo.Nodes;
 using Dynamo.Utilities;
+using Dynamo.Logging;
 using ProtoCore.Namespace;
 
 namespace Dynamo.Models
