+++ src/DynamoCore/Nodes/FunctionCallNodeController.cs
@@ -5,6 +5,7 @@ using System.Xml;
 using Dynamo.DSEngine;
 using Dynamo.Interfaces;
 using Dynamo.Models;
+using Dynamo.Logging;
 
 using ProtoCore.AST.AssociativeAST;
 using ProtoCore.Lang;
